fix: Remove json-format plugin (#552)
This plugin has been a source of multiple problems so we are removing it.
Also, since we removed it, reading JSON files when the parser does not support returns an error. So we'll filter it for now.
Added a test to check if we can analyze JSON and JS files when using typescript patterns.