fix: Fix and test all-patterns for typescript files
Since the @typescript-eslint plugin requires the tsconfig.json file and
some other settings to run correctly, we have some special code here:
https://github.com/codacy/codacy-eslint/blob/768b63202b11ab76610ff88e09edfab861704393/src/configCreator.ts#L51
that disables the typescript patterns if not all the files are
typescript files. This made the last test with a mixed folder with js
and ts files ineffectful to test the typescript part.
Creating a different test with only typescript files showed another missing
default.