change: add debug messages to help when DEBUG=1 passed
This allow us to run the docker tool locally to try to
understand what's going on with the it.
Covers the main important points, specially what exactly
are the eslint settings being applied for each file since
eslint configuration resolution can be quite complex.

All the debug* functions are protected by verifying that DEBUG=1
so that none of the special code for debugging is run under normal
conditions. Just to make sure we don't degrade performance.

The docker can be run with:

docker run -e DEBUG=1 -v (pwd):/src:ro -v (pwd)/.codacyrc:/.codacyrc -it --rm codacy-eslint:latest

for example.